style: move scratch to .github
