+++ .github/scratch/paper/index.js
@@ -1,4 +1,4 @@
-import { List, AppendList, utils } from '../../index.js'
+import { List, AppendList, utils } from '../../../index.js'
 
 class ViewPosition {
   constructor({ x = 0, y = 0 } = {}) {
