+++ package.json
@@ -11,7 +11,7 @@
     "test:coverage": "NODE_ENV=test node --test --experimental-test-coverage test/index.spec.js",
 
     "lint": "npx eslint",
-    "scratch": "node playground/index.js",
+    "scratch": "node .github/scratch/index.js",
     "example": "node .github/example/index.js"
   },
   "engines": {
